Restart unless stopped (#517)
Restarting unless-stopped is a good practice in production-like
environments (and most of the non-production systems).

It improves resilience (containers will restart if node is rebooted or
if one of the containers gets killed by OOM); it also helps with
predictibility as it won't restart the containers if we intentionally
stopped them.